+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- Raise with an explicit message when all or find are given nil.
+
 ## 3.0.1
 
 - Fix doc typo